+++ pdns/shuffle.hh
@@ -20,32 +20,14 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 #pragma once
-#include <cstdint>
-#include <limits>
+#include <vector>
 
-void dns_random_init(const std::string& data = "", bool force_reinit = false);
-uint32_t dns_random(uint32_t n);
-uint16_t dns_random_uint16();
+struct DNSRecord;
+struct DNSZoneRecord;
 
 namespace pdns {
-  struct dns_random_engine {
-
-    typedef uint32_t result_type;
-
-    static constexpr result_type min()
-    {
-      return 0;
-    }
-
-    static constexpr result_type max()
-    {
-      return std::numeric_limits<result_type>::max() - 1;
-    }
-
-    result_type operator()()
-    {
-      return dns_random(std::numeric_limits<result_type>::max());
-    }
-  };
+  void shuffle(std::vector<DNSRecord>& rrs);
+  void shuffle(std::vector<DNSZoneRecord>& rrs);
+  void orderAndShuffle(std::vector<DNSRecord>& rrs);
 }
